+++ app/Filament/Resources/RatingMorphResource/Schemas/RatingMorphInfolist.php
@@ -13,7 +13,11 @@ class RatingMorphInfolist extends XotBaseResourceInfolist
     /**
      * @return array<string, Component>
      */
+<<<<<<< HEAD
     public function getInfolistSchema(): array
+=======
+    public static function getInfolistSchema(): array
+>>>>>>> e8cf105 (Check & fix styling)
     {
         return [
             'model_id' => TextEntry::make('model_id'),