+++ app/Filament/Resources/RatingResource.php
@@ -4,12 +4,16 @@ declare(strict_types=1);
 
 namespace Modules\Rating\Filament\Resources;
 
+<<<<<<< HEAD
 use Modules\Rating\Filament\Resources\RatingResource\RelationManagers\ChildrenRelationManager;
+=======
+>>>>>>> e8cf105 (Check & fix styling)
 use Modules\Rating\Models\Rating;
 
 class RatingResource extends BaseRatingResource
 {
     protected static ?string $model = Rating::class;
+<<<<<<< HEAD
 
     /**
      * @return array<int, class-string>
@@ -20,4 +24,6 @@ class RatingResource extends BaseRatingResource
             ChildrenRelationManager::class,
         ];
     }
+=======
+>>>>>>> e8cf105 (Check & fix styling)
 }