+++ app/Models/Contracts/RatingContract.php
@@ -4,14 +4,7 @@ declare(strict_types=1);
 
 namespace Modules\Rating\Models\Contracts;
 
-<<<<<<< HEAD
-<<<<<<< HEAD
 use Illuminate\Database\Eloquent\Collection;
-=======
->>>>>>> fd7a600 (.)
-=======
-use Illuminate\Database\Eloquent\Collection;
->>>>>>> laraxot/dev
 use Illuminate\Database\Eloquent\Model;
 use Modules\Rating\Enums\RuleEnum;
 use Modules\Xot\Contracts\HasRecursiveRelationshipsContract;
@@ -45,10 +38,6 @@ use Modules\Xot\Contracts\HasRecursiveRelationshipsContract;
  *
  * @phpstan-require-extends Model
  *
-<<<<<<< HEAD
-<<<<<<< HEAD
-=======
->>>>>>> laraxot/dev
  * @property int                    $id
  * @property int|null               $parent_id
  * @property string|null            $title
@@ -59,20 +48,6 @@ use Modules\Xot\Contracts\HasRecursiveRelationshipsContract;
  * @property int|null               $order_column
  * @property Collection<int, Model> $children
  * @property RuleEnum               $rule
-<<<<<<< HEAD
-=======
- * @property int         $id
- * @property int|null    $parent_id
- * @property string|null $title
- * @property string|null $txt
- * @property string|null $slug
- * @property bool|null   $is_readonly
- * @property bool|null   $is_disabled
- * @property int|null    $order_column
- * @property RuleEnum    $rule
->>>>>>> fd7a600 (.)
-=======
->>>>>>> laraxot/dev
  */
 interface RatingContract extends HasRecursiveRelationshipsContract
 {