+++ app/Models/Traits/HasLikes.php
@@ -9,33 +9,66 @@ use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Modules\Rating\Models\Like;
 use Modules\Xot\Contracts\UserContract;
 
+<<<<<<< HEAD
 trait HasLikes
 {
     /**
      * @return Collection<int, Like>
      */
     public function likes(): Collection
+=======
+/** @phpstan-ignore trait.unused */
+trait HasLikes
+{
+    /**
+     * @return Collection
+     */
+    public function likes()
+>>>>>>> e8cf105 (Check & fix styling)
     {
         return $this->likesRelation;
     }
 
+<<<<<<< HEAD
     public function likedBy(?UserContract $user): void
     {
         if (null === $user) {
             return;
         }
 
+=======
+    /**
+     * param \Modules\Xot\Contracts\UserContract|null $user.
+     *
+     * @param UserContract|null $user
+     */
+    public function likedBy($user): void
+    {
+>>>>>>> e8cf105 (Check & fix styling)
         $this->likesRelation()->create(['user_id' => $user->id]);
 
         $this->unsetRelation('likesRelation');
     }
 
+<<<<<<< HEAD
     public function dislikedBy(?UserContract $user): void
     {
         if (null === $user) {
             return;
         }
 
+=======
+    /**
+     * param \Modules\Xot\Contracts\UserContract|null $user.
+     *
+     * @param UserContract|null $user
+     */
+    public function dislikedBy($user): void
+    {
+        /**
+         * @var Like
+         */
+>>>>>>> e8cf105 (Check & fix styling)
         $where = $this->likesRelation()->where('user_id', $user->id)->first();
         if (null !== $where) {
             $where->delete();
@@ -50,12 +83,16 @@ trait HasLikes
      *
      * @see https://github.com/laravelio/laravel.io/issues/350
      */
+<<<<<<< HEAD
     /** @return MorphMany<Like, $this> */
+=======
+>>>>>>> e8cf105 (Check & fix styling)
     public function likesRelation(): MorphMany
     {
         return $this->morphMany(Like::class, 'likesRelation', 'likeable_type', 'likeable_id');
     }
 
+<<<<<<< HEAD
     public function isLikedBy(?UserContract $user): bool
     {
         if (null === $user) {
@@ -69,6 +106,29 @@ trait HasLikes
     {
         static::deleting(static function (self $model): void {
             $model->likesRelation()->delete();
+=======
+    /**
+     * param \Modules\Xot\Contracts\UserContract|null $user.
+     *
+     * @param UserContract|null $user
+     *
+     * @return bool
+     */
+    public function isLikedBy($user)
+    {
+        return $this->likesRelation()->where('user_id', $user->id)->exists();
+    }
+
+    /**
+     * Undocumented function.
+     *
+     * @return void
+     */
+    protected static function bootHasLikes()
+    {
+        static::deleting(function ($model): void {
+            $model->likesRelation()->delete(); /* @phpstan-ignore method.nonObject */
+>>>>>>> e8cf105 (Check & fix styling)
             $model->unsetRelation('likesRelation');
         });
     }