+++ docs/changelog.md
@@ -1,3 +1,4 @@
+<<<<<<< HEAD
 ---
 module: theme
 topic: CHANGELOG
@@ -5,3 +6,12 @@ canonical: ../../../Themes/docs/shared-components/CHANGELOG.md
 ---
 
 See canonical documentation: ../../../Themes/docs/shared-components/CHANGELOG.md
+=======
+# Changelog
+
+All notable changes to `:package_name` will be documented in this file.
+
+## 1.0.0 - 202X-XX-XX
+
+- initial release
+>>>>>>> e8cf105 (Check & fix styling)